create more factories from data-catelog, add onClickTitle
